feat: remove useEarnBalances (#3809)
