+++ src/pages/Dashboard/PortfolioBreakdown.tsx
@@ -6,8 +6,10 @@ import { Card } from 'components/Card/Card'
 import { CircularProgress } from 'components/CircularProgress/CircularProgress'
 import { Text } from 'components/Text'
 import { bnOrZero } from 'lib/bignumber/bignumber'
-import { useEarnBalances } from 'pages/Defi/hooks/useEarnBalances'
-import { selectPortfolioTotalFiatBalanceExcludeEarnDupes } from 'state/slices/selectors'
+import {
+  selectEarnBalancesFiatAmountFull,
+  selectPortfolioTotalFiatBalanceExcludeEarnDupes,
+} from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
 type StatCardProps = {
@@ -49,16 +51,11 @@ const BreakdownCard: React.FC<StatCardProps> = ({
 
 export const PortfolioBreakdown = () => {
   const history = useHistory()
-  //FOXY, OSMO, COSMO, Yarn Vaults
-  // TODO(gomes): This goes away in a follow-up PR
-  // - FOXy balances are now the only effective reason we have a useEarnBalances( hook, and a selector should be able to get that
-  // - Once useEarnBalances() is removed, we should be able to properly get earn balances from selector, meaning the total balance will accurately be
-  // the same as the addition below
-  const earnBalances = useEarnBalances()
+  const earnBalance = useAppSelector(selectEarnBalancesFiatAmountFull).toFixed()
   const portfolioTotalFiatBalance = useAppSelector(selectPortfolioTotalFiatBalanceExcludeEarnDupes)
   const netWorth = useMemo(
-    () => bnOrZero(earnBalances.totalEarningBalance).plus(portfolioTotalFiatBalance).toFixed(),
-    [earnBalances.totalEarningBalance, portfolioTotalFiatBalance],
+    () => bnOrZero(earnBalance).plus(portfolioTotalFiatBalance).toFixed(),
+    [earnBalance, portfolioTotalFiatBalance],
   )
   return (
     <Flex gap={{ base: 0, xl: 6 }} flexDir={{ base: 'column', md: 'row' }}>
@@ -67,15 +64,13 @@ export const PortfolioBreakdown = () => {
         percentage={bnOrZero(portfolioTotalFiatBalance).div(netWorth).times(100).toNumber()}
         label='defi.walletBalance'
         onClick={() => history.push('/accounts')}
-        isLoading={earnBalances.loading}
       />
       <BreakdownCard
-        value={earnBalances.totalEarningBalance}
-        percentage={bnOrZero(earnBalances.totalEarningBalance).div(netWorth).times(100).toNumber()}
+        value={earnBalance}
+        percentage={bnOrZero(earnBalance).div(netWorth).times(100).toNumber()}
         label='defi.earnBalance'
         color='green.500'
         onClick={() => history.push('/defi')}
-        isLoading={earnBalances.loading}
       />
     </Flex>
   )
